pyproject.toml
[tool.poetry]
name = "fastapi-celery-app"
version = "0.1.0"
description = ""
authors = ["solarized-penguin <solarized.penguin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = {extras = ["redis"], version = "^5.4.0"}
fastapi = "^0.111.0"
pydantic = {extras = ["all"], version = "^2.7.3"}
sqlmodel = {extras = ["postgres"], version = "^0.0.19"}
flower = "^2.0.1"
pydantic-settings = "^2.3.1"
python-logging-loki = "^0.3.1"
fastapi-mail = {extras = ["httpx"], version = "^1.4.1"}
pymongo = {extras = ["encryption"], version = "^4.7.3"}
beanie = {extras = ["queue"], version = "^1.26.0"}


[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"
black = {extras = ["d"], version = "^24.4.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
